+++ mono_repo/test/mono_repo_test.dart
@@ -29,7 +29,8 @@ void main() {
   });
 }
 
-const _helpOutput = '''Manage multiple packages in one source repository.
+const _helpOutput = r'''
+Manage multiple packages in one source repository.
 
 Usage: mono_repo <command> [arguments]
 
@@ -41,8 +42,9 @@ Global options:
 
 Available commands:
   check       Check the state of the repository.
-  presubmit   Run the travis presubmits locally.
+  generate    Generates the CI configuration for child packages.
+  presubmit   Run the ci presubmits locally.
   pub         Run `pub get` or `pub upgrade` against all packages.
-  travis      Configure Travis-CI for child packages.
+  travis      (Deprecated, use `generate`) Configure Travis-CI for child packages.
 
 Run "mono_repo help <command>" for more information about a command.''';