+++ mono_repo/test/shared.dart
@@ -7,7 +7,8 @@ import 'dart:async';
 import 'package:checked_yaml/checked_yaml.dart';
 import 'package:io/ansi.dart';
 import 'package:mono_repo/src/ci_shared.dart';
-import 'package:mono_repo/src/commands/travis.dart';
+import 'package:mono_repo/src/commands/ci_script/generate.dart';
+import 'package:mono_repo/src/commands/travis/generate.dart';
 import 'package:mono_repo/src/root_config.dart';
 import 'package:mono_repo/src/user_exception.dart';
 import 'package:path/path.dart' as p;
@@ -28,10 +29,15 @@ void testGenerateTravisConfig({
         })).run(
       () => overrideAnsiOutput(
         false,
-        () => generateTravisConfig(
-          RootConfig(rootDirectory: d.sandbox),
-          validateOnly: validateOnly,
-        ),
+        () {
+          final config = RootConfig(rootDirectory: d.sandbox);
+          logPackages(config);
+          generateTravisConfig(
+            config,
+            validateOnly: validateOnly,
+          );
+          generateCIScript(config, validateOnly: validateOnly);
+        },
       ),
     );
   } finally {
@@ -88,6 +94,6 @@ stages:
     - test: --preset travis
 ''';
 
-String get travisShPathMessage => '''
-${scriptLines(travisShPath).join('\n')}
-Wrote `${p.join(d.sandbox, travisShPath)}`.''';
+String get ciScriptPathMessage => '''
+${scriptLines(ciScriptPath).join('\n')}
+Wrote `${p.join(d.sandbox, ciScriptPath)}`.''';